requirements.txt
aiocron==1.8
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
croniter==3.0.3
deep-translator==1.11.4
fake-useragent==1.5.1
ffmpeg==1.4
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
idna==3.8
magic-filter==1.0.12
multidict==6.0.5
numpy==2.1.0
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
Telethon==1.36.0
termcolor==2.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
